requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-heroku==0.0.0
djangorestframework==3.12.2
gunicorn==20.0.4
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.0
pytesseract==0.3.7
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
